Add support for more modern (non-deprecated) table view swipe API
UITableViewRowAction and it’s associated methods have been deprecated in favor of the new UISwipeActionsConfiguration object which can be configured on the trailing and leading edges of a table view cell.

This API is available on iOS 11.0 onwards. The cleanest way to support this is to bump the deployment target to iOS 11.